chore(fe): modify assignment info tab ui
